Restructured js "proxies" into separate objects
