+++ icecandidate.js
@@ -1,0 +1,7 @@
+function RTCIceCandidate(dict) {
+  this.candidate = dict.candidate;
+  this.sdpMid = dict.sdpMid;
+  this.sdpMLineIndex = dict.hasOwnProperty('sdpMLineIndex') ? dict.sdpMLineIndex : 0;
+}
+
+module.exports = RTCIceCandidate;
