+++ sessiondescription.js
@@ -1,0 +1,6 @@
+function RTCSessionDescription(dict) {
+  this.type = dict.type;
+  this.sdp = dict.sdp;
+};
+
+module.exports = RTCSessionDescription;
